refactor(store): extract _IndexBackendMixin to _index.py (#280) (#379)
Third step of the VstashStore decomposition (#280). Move-only: the 11
vector-index methods (_snapvec_path, _ivfpq_path, _init_ivfpq, _nlist_for,
_derive_nlist, fit_ivfpq, _init_snapvec, _rebuild_snapvec_from_vec_chunks,
_save_snapvec, _checkpoint_snapvec, _reload_snapvec) move verbatim into
_IndexBackendMixin in vstash/store/_index.py; VstashStore inherits it.

To avoid an _index <-> __init__ import cycle, the snapvec capability probe
(SnapIndex / _HAS_SNAPVEC) moves to the package leaf _common.py, with a
SnapIndex=None fallback when snapvec is absent (every use site is guarded by
_HAS_SNAPVEC). __init__.py and _index.py both import them from _common; the
_HAS_SNAPVEC re-export from vstash.store is preserved.

Method bodies are byte-identical (verified via diff). Class-level annotations on
the mixin declare the host-provided attributes (carried over from the review
feedback on the schema mixin).

- pytest tests/ -> 1251 passed (1248 + 3 new test_index_backend.py), 6 deselected
- ruff check/format clean; MRO: VstashStore -> _IndexBackendMixin -> _SchemaManagerMixin
